UIFC-370-check-urls-of-showAllCollections-and-showSelectedCollections outsource doShowCollections, add test
